move RuntimeUpdateStrategy one package up
