+++ openwebstart/src/main/java/com/openwebstart/jvm/JavaRuntimeSelector.java
@@ -3,7 +3,6 @@ package com.openwebstart.jvm;
 import com.openwebstart.http.DownloadInputStream;
 import com.openwebstart.jvm.runtimes.LocalJavaRuntime;
 import com.openwebstart.jvm.runtimes.RemoteJavaRuntime;
-import com.openwebstart.jvm.runtimes.RuntimeUpdateStrategy;
 import com.openwebstart.jvm.util.RuntimeVersionComparator;
 import net.adoptopenjdk.icedteaweb.Assert;
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionString;
