+++ openwebstart/src/main/java/com/openwebstart/jvm/RuntimeManagerConfig.java
@@ -1,6 +1,5 @@
 package com.openwebstart.jvm;
 
-import com.openwebstart.jvm.runtimes.RuntimeUpdateStrategy;
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionString;
 
 import java.net.URI;
