+++ openwebstart/src/main/java/com/openwebstart/jvm/RuntimeUpdateStrategy.java
@@ -1,4 +1,4 @@
-package com.openwebstart.jvm.runtimes;
+package com.openwebstart.jvm;
 
 public enum RuntimeUpdateStrategy {
